requirements-build.txt
jinja2==3.1.6
jinja2-workarounds==0.1.0
markdown==3.10.2
markupsafe==3.0.3  # not semver, be cautious
mkdocs-glightbox==0.5.2
mkdocs-git-revision-date-localized-plugin==1.5.1
mkdocs-material==9.7.5
mkdocs-redirects==1.2.2
mkdocs-table-reader-plugin==3.1.0
pymdown-extensions==10.21
git+https://github.com/wwarriner/mkdocs-macros-plugin@d445c98dcc14ecae48ddf53b4758f3942ed706b8
git+https://github.com/uabrc/mkdocs@888fd28f92a320352f63600c24635231a42e5fac
git+https://github.com/uabrc/mkdocs-title-casing-plugin.git@stable
